+++ src/app/beer_garden/config.py
@@ -1494,6 +1494,13 @@ _PLUGIN_SPEC = {
                         "to local plugin processes"
                     ),
                 },
+                "capture_streams": {
+                    "type": "bool",
+                    "default": False,
+                    "description": (
+                        "Capture STDOUT and STDERR streams for local plugins and output to file"
+                    ),
+                },
                 "logging": {
                     "type": "dict",
                     "items": {